requirements.txt
alabaster==0.7.12
aniso8601==3.0.2
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
Django==1.11.7
django-filter==1.1.0
django-graphql-jwt==0.1.14
docutils==0.14
gevent==1.3.7
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
greenlet==0.4.15
gunicorn==19.7.1
idna==2.8
imagesize==1.1.0
ipython-genutils==0.2.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
nbformat==4.4.0
networkx==2.2
numpy==1.15.4
packaging==18.0
pandas==0.23.4
Pillow==5.3.0
plotly==3.6.0
promise==2.2.1
psycopg2==2.7.3.2
Pygments==2.3.1
PyJWT==1.6.4
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.21.0
retrying==1.3.3
Rx==1.6.1
scikit-learn==0.20.3
scipy==1.2.0
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
